server/requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.7.0
cached-property==1.5.1
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.5.0
distro-info===0.23ubuntu1
Django==3.1.1
django-appconf==1.0.4
django-bootstrap-toolkit==2.15.0
django-bootstrap3==14.1.0
django-compressor==2.4
django-cors-headers==3.5.0
django-denorm==1.0.0
django-extensions==2.2.9
django-guardian==2.3.0
django-http-proxy==0.4.3
django-reversion==3.0.8
django-select2==7.4.2
django-userena-ce==6.0.1
djangorestframework==3.11.1
docker==4.3.1
docker-compose==1.27.3
dockerpty==0.4.1
docopt==0.6.2
easy-thumbnails==2.7
entrypoints==0.3
feedgen==0.9.0
filelock==3.0.12
GDAL==3.0.4
geojson==2.5.0
gunicorn==20.0.4
html2text==2020.1.16
httplib2==0.14.0
idna==2.8
isort==5.5.3
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.12.0
lxml==4.5.2
macaroonbakery==1.3.1
mccabe==0.6.1
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
ordereddict==1.1
paramiko==2.7.2
Paver==1.3.4
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
pyrsistent==0.17.3
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.14.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
rcssmin==1.0.6
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rjsmin==1.1.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.3.1
systemd-python==234
texttable==1.6.3
toml==0.10.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.0.31
wadllib==1.3.3
websocket-client==0.57.0
wrapt==1.12.1
xkit==0.0.0
